+++ packages/metro-config/test/index.test.ts
@@ -1,6 +1,3 @@
-// @ts-check
-"use strict";
-
 describe("@rnx-kit/metro-config", () => {
   const fs = require("fs");
   const path = require("path");
@@ -28,17 +25,15 @@ describe("@rnx-kit/metro-config", () => {
 
   /**
    * Returns path to specified test fixture.
-   * @param {string} name
    */
-  function fixturePath(name) {
+  function fixturePath(name: string): string {
     return path.join(currentWorkingDir, "test", "__fixtures__", name);
   }
 
   /**
    * Sets current working directory to specified test fixture.
-   * @param {string} name
    */
-  function setFixture(name) {
+  function setFixture(name: string): void {
     process.chdir(fixturePath(name));
   }
 
@@ -255,6 +250,8 @@ describe("@rnx-kit/metro-config", () => {
       fail("Expected `config.resolver.blockList` to be a RegExp");
     } else if (!config.transformer) {
       fail("Expected `config.transformer` to be defined");
+    } else if (!config.transformer.getTransformOptions) {
+      fail("Expected `config.transformer.getTransformOptions` to be defined");
     } else if (!Array.isArray(config.watchFolders)) {
       fail("Expected `config.watchFolders` to be an array");
     }
@@ -267,8 +264,13 @@ describe("@rnx-kit/metro-config", () => {
     expect(config.resolver.blacklistRE.source).toBe(blockList);
     expect(config.resolver.blockList.source).toBe(blockList);
 
-    const transformerOptions = await config.transformer.getTransformOptions();
-    expect(transformerOptions.transform).toEqual({
+    const opts = { dev: false, hot: false };
+    const transformerOptions = await config.transformer.getTransformOptions(
+      [],
+      opts,
+      () => Promise.resolve([])
+    );
+    expect(transformerOptions?.transform).toEqual({
       experimentalImportSupport: false,
       inlineRequires: false,
     });
@@ -299,6 +301,8 @@ describe("@rnx-kit/metro-config", () => {
       fail("Expected `config.resolver.blockList` to be a RegExp");
     } else if (!config.transformer) {
       fail("Expected `config.transformer` to be defined");
+    } else if (!config.transformer.getTransformOptions) {
+      fail("Expected `config.transformer.getTransformOptions` to be defined");
     } else if (!Array.isArray(config.watchFolders)) {
       fail("Expected `config.watchFolders` to be an array");
     }
@@ -311,8 +315,13 @@ describe("@rnx-kit/metro-config", () => {
     expect(config.resolver.blacklistRE.source).toBe(blockList);
     expect(config.resolver.blockList.source).toBe(blockList);
 
-    const transformerOptions = await config.transformer.getTransformOptions();
-    expect(transformerOptions.transform).toEqual({
+    const opts = { dev: false, hot: false };
+    const transformerOptions = await config.transformer.getTransformOptions(
+      [],
+      opts,
+      () => Promise.resolve([])
+    );
+    expect(transformerOptions?.transform).toEqual({
       experimentalImportSupport: false,
       inlineRequires: false,
     });