+++ packages/test-app/CHANGELOG.md
@@ -1,8 +1,40 @@
 # Change Log - @rnx-kit/test-app
 
-This log was last generated on Thu, 26 Aug 2021 17:24:29 GMT and should not be manually modified.
+This log was last generated on Tue, 31 Aug 2021 10:50:41 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.0.1
+
+Tue, 31 Aug 2021 10:50:41 GMT
+
+### Patches
+
+- Bump @rnx-kit/test-app to v0.0.1 (4123478+tido64@users.noreply.github.com)
+
+## 0.0.1
+
+Tue, 31 Aug 2021 06:43:13 GMT
+
+### Patches
+
+- Bump @rnx-kit/test-app to v0.0.1 (4123478+tido64@users.noreply.github.com)
+
+## 0.0.1
+
+Tue, 31 Aug 2021 06:21:44 GMT
+
+### Patches
+
+- Bump @rnx-kit/test-app to v0.0.1 (4123478+tido64@users.noreply.github.com)
+
+## 0.0.1
+
+Fri, 27 Aug 2021 18:41:43 GMT
+
+### Patches
+
+- Bump @rnx-kit/test-app to v0.0.1 (4123478+tido64@users.noreply.github.com)
 
 ## 0.0.1