+++ packages/typescript-service/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @rnx-kit/typescript-service
 
-This log was last generated on Wed, 25 Aug 2021 08:52:48 GMT and should not be manually modified.
+This log was last generated on Tue, 31 Aug 2021 06:43:13 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.3.4
+
+Tue, 31 Aug 2021 06:43:13 GMT
+
+### Patches
+
+- Bump @rnx-kit/typescript-service to v1.3.4 (4123478+tido64@users.noreply.github.com)
+
+## 1.3.3
+
+Fri, 27 Aug 2021 18:41:43 GMT
+
+### Patches
+
+- Bump @rnx-kit/typescript-service to v1.3.3 (4123478+tido64@users.noreply.github.com)
 
 ## 1.3.2